*add shit to the noHDD branch* this is based on r877 or so with muchos changes-o.  The HDD detection has been moved to a idle thread and the app can start without a usb drive.  The gamelist will be populated with titles until a USB drive is detected.  Then the gamelist should reload itself with wii games.
still todo...
im not 100% sure if i broke campatability for some drives, or increased
it for others.  i added a USB_deinit() before most of the USB_init() and
it works for me and thats what matters.

the shutdown and reset callbacks work fine on this thread, so they can
be removed from all loops where the thread is still running.

the checkthread needs to be halted when removing elements from the
screen.  ive done it for most of the prompts and shit, but you can still
see the ugliness when changing pages in the settings with no hdd
attached.

the sd detection in the checkthread only detects when the SD card is
removed and not when reinserted.

there is no way to format a drive as-is.  i have set up MENU_FORMAT on a
button in the settings, but some of the shit from MENU_CHECK needs to be
added to it before it will function.